+++ packages/core/src/generators/qwik/helpers/convert-method-to-function.ts
