+++ packages/core/src/generators/qwik/helpers/stable-inject.ts
