+++ packages/core/src/generators/qwik/helpers/stable-serialize.ts
